style(oauth-jwt): Rename the classes
