+++ src/main/java/com/chatcode/config/auth/oauth/dto/OAuth2Response.java
@@ -1,18 +1,18 @@
-package com.chatcode.config.auth.oauth;
+package com.chatcode.config.auth.oauth.dto;
 
 import java.util.Map;
 import lombok.Builder;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 
 @Builder
-public record OAuth2UserInfo(
+public record OAuth2Response(
         String name,
         String email,
         String profile,
         String providerId,
         String provider
 ) {
-    public static OAuth2UserInfo of(String registrationId, Map<String, Object> attributes)
+    public static OAuth2Response of(String registrationId, Map<String, Object> attributes)
             throws OAuth2AuthenticationException {
         return switch (registrationId) {
             case "google" -> ofGoogle(attributes);
@@ -21,8 +21,8 @@ public record OAuth2UserInfo(
         };
     }
 
-    private static OAuth2UserInfo ofGoogle(Map<String, Object> attributes) {
-        return OAuth2UserInfo.builder()
+    private static OAuth2Response ofGoogle(Map<String, Object> attributes) {
+        return OAuth2Response.builder()
                 .name((String) attributes.get("name"))
                 .email((String) attributes.get("email"))
                 .profile((String) attributes.get("picture"))
@@ -31,8 +31,8 @@ public record OAuth2UserInfo(
                 .build();
     }
 
-    private static OAuth2UserInfo ofGithub(Map<String, Object> attributes) {
-        return OAuth2UserInfo.builder()
+    private static OAuth2Response ofGithub(Map<String, Object> attributes) {
+        return OAuth2Response.builder()
                 .name((String) attributes.get("login"))
                 .email((String) attributes.get("email"))
                 .profile((String) attributes.get("avatar_url"))
